+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed crash on servers
+- Fixed integer overflow on some tasks